% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{character.matrix}
\alias{character.matrix}
\title{Character matrix; 2 columns of 10 letters.}
\format{Unnamed matrix with 20 characters}
\usage{
character.matrix
}
\description{
Character matrix; 2 columns of 10 letters.
}
\keyword{datasets}
