% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartcolors.R
\name{ChartColors}
\alias{ChartColors}
\title{Generates a vector of colors}
\usage{
ChartColors(
  number.colors.needed,
  given.colors,
  reverse = FALSE,
  palette.start = 0,
  palette.end = 1,
  trim.light.colors = TRUE,
  custom.color = NA,
  custom.gradient.start = NA,
  custom.gradient.end = NA,
  custom.palette = NA,
  silent = FALSE,
  silent.single.color = silent
)
}
\arguments{
\item{number.colors.needed}{The number of colors to generate.}

\item{given.colors}{Specifies the color vector to output. It can be;
(1) A named palette from grDevices, RColorBrewer colorspace, or colorRamps;
(2) A vector of colors which will be recycled to length \code{number.colors.needed}; or
(3) one of \code{"Custom color"}, \code{"Custom gradient"} or \code{"Custom palette"}.
The last option gives the user greater control via additional parameters (see below).  If not specified, the colors used
are c("#5C9AD3", "#ED7D31", "#A5A5A5", "#FFC000", "#4473C5", "#70AD46",
            "#255F91", "#9E480D", "#636365", "#987300", "#26408B", "#42682B")}

\item{reverse}{Whether the output color vector should be reversed. Ignored when \code{given.colors}
is any of \code{"Custom color"}, \code{"Custom gradient"} or \code{"Custom palette"}.}

\item{palette.start}{A numeric in [0,1] specifying the start of the palette relative to the full spectrum
of \code{given.colors}. The default is 0, which means the palette starts with the first color of \code{given.colors}.}

\item{palette.end}{A numeric in [0,1] specifying the end of the palette relative to the full spectrum
of \code{given.colors}. The default is 1, which means the palette starts with the last color of \code{given.colors}.}

\item{trim.light.colors}{When selected, palette.start and palette.end will be set to remove light colors in the
monochrome palettes (\code{"Blues","Greens","Greys","Oranges","Purples","Reds"}), if this is not already achieved
by \code{palette.start} and \code{palette.end}.}

\item{custom.color}{A single color provided as a hex or character string. Only used if \code{given.colors}
is \code{"Custom color"}. The output vector will consist of \code{custom.color} repeated \code{number.colors.needed}
(no interpolation).}

\item{custom.gradient.start}{A color specifying the start of the gradient when \code{given.colors}
is set to \code{"Custom gradient"}.}

\item{custom.gradient.end}{A color specifying the end of the gradient when \code{given.colors}
is set to \code{"Custom gradient"}.}

\item{custom.palette}{A vector or comma separated list of colors which will be recycled to the desired length.
Only used if \code{given.colors} is \code{"Custom palette"}.}

\item{silent}{Option to hide warnings about the number of colors.}

\item{silent.single.color}{Option to hide warnings if only a single color is provided when 
\code{number.colors.needed > 1}. This type of warning is controlled separately from the \code{silent} because
there are some chart types for which a single color or a palette can be appropriately specified. In contrast
the times when colors need to recycled for fill the palette, or the a palette is used for a single color
usually indicates the user has specified some inappropriate options.}
}
\description{
Generates a vector of colors from a palette or vector of colors.
Alpha channels are ignored and set to 255/FF.
}
\examples{
ChartColors(5, given.colors = c("blue", "orange", "green"))
ChartColors(5, given.colors = "blue")
ChartColors(5, given.colors = "#9CFF73")
ChartColors(5, given.colors = "Set3", reverse = TRUE)
plot(1:10,1:10,pch=19, col=ChartColors(10,"Reds",trim.light.colors=TRUE, reverse=TRUE))
plot(1:12, 1:12, pch=19, col=ChartColors(12))
}
