% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartnumberformat.R
\name{ChartNumberFormat}
\alias{ChartNumberFormat}
\title{Generates a d3 number format for use by charting functions.}
\usage{
ChartNumberFormat(number.format)
}
\arguments{
\item{number.format}{A list of five unnamed items in the following order:
1) The type of number formatting. One of \code{"Number"}, \code{"Percentage"}, \code{"Date/Time"},
 \code{"Currency"}, \code{"Metric units suffix"}, \code{"Scientific"} and \code{"Custom"}.
2) A d3 date format. If specified, this will be used provided a custom format is not specified.
3) A custom d3 number format. If specifed this will be used in prefernce to all other inputs.
4) A boolean indicating whether to separate thousands with commas.
5) An integer specifying the number of decimal places, or for \code{"Metric units suffix"}, the number
of significant digits.}
}
\description{
See https://github.com/d3/d3-format for more information on d3.
}
\details{
Defaults to number with zero decimal places and no thousand separation if no type,
date format or custom format is specified.
}
