% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnumcolors.R
\name{GetNumColors}
\alias{GetNumColors}
\title{Determines the number of colors required in a chart}
\usage{
GetNumColors(data, chart.type)
}
\arguments{
\item{data}{Input data, which is assumed to be created by \code{\link{PrepareData}}}

\item{chart.type}{Type of chart to plot}
}
\description{
Returns the number of colors needed in a chart.
This is returned as a list because more than one color-series may be required (e.g. Pie chart)
}
