% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getvector.R
\name{GetVector}
\alias{GetVector}
\title{GetVector}
\usage{
GetVector(var.names, length = NA, value = NA)
}
\arguments{
\item{var.names}{The name of a variable, or a vector of variable names as characters.}

\item{length}{The desired length of the vector.}

\item{value}{The default value of the vector.}
}
\value{
A vector, or list of vectors. If the variable does not exist or is \code{NULL}, the return value will be a vector of form \code{rep(value, length)}.
}
\description{
A convenience function that will return a vector of a specified length
}
\examples{
a <- 1
GetVector(c("a","b", "c"), 5, 0)
}
