% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscales.R
\name{MapToColors}
\alias{MapToColors}
\title{Construct color scale for a numeric value}
\usage{
MapToColors(
  x,
  min.x = if (length(x) == 1) 1 else min(x),
  max.x = max(x),
  mid.x = NULL,
  min.color = "white",
  mid.color = "Displayr grey",
  max.color = "Displayr blue"
)
}
\arguments{
\item{x}{A numeric vector specifying the colors returne}

\item{min.x}{A numeric value representing the smallest value on the color scale. Values in 
\code{x} smaller than this will be treated as \code{min.x}.}

\item{max.x}{A numeric value representing the largest value on the color scale. Values in 
\code{x} larger than this will be treated as \code{max.x}.}

\item{mid.x}{A numeric value representing the midpoint on the color scale.}

\item{min.color}{The color representing the smallest value.}

\item{mid.color}{The color representing the mid point on the color scale.}

\item{max.color}{The color representing the max point on the color scale.}
}
\description{
Construct color scale for a numeric value
}
