% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchtable.R
\name{MatchTable}
\alias{MatchTable}
\title{Match table to elements in the reference table}
\usage{
MatchTable(x, ref.table = NULL, ref.maindim = "rows",
  ref.names = NULL, x.table.name = "", ref.table.name = "input data",
  as.matrix = TRUE, trim.whitespace = TRUE,
  silent.remove.duplicates = FALSE)
}
\arguments{
\item{x}{A vector or table containing values to be matched against \code{ref.table}}

\item{ref.table}{The reference table which determines the order of the output vector}

\item{ref.maindim}{Row or column.}

\item{ref.names}{Alternatively, a vector of names can be provide instead of \code{ref.table} and \code{ref.maindim}.}

\item{x.table.name}{Name used in the error messages.}

\item{ref.table.name}{Name used in the error messages.}

\item{as.matrix}{Converts \code{x} to a matrix. This will force all values to be of the same type.}

\item{trim.whitespace}{Whether to trim leading and trailing whitespace when matching row or column names.}

\item{silent.remove.duplicates}{Removes duplicates with giving warnings. This is particulary useful when dealing with banners.}
}
\description{
Match table to elements in the reference table
}
