% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparecolors.R
\name{PrepareColors}
\alias{PrepareColors}
\title{Generates vectors of colors for use by CChart.
Wrapper for \code{GetNumColors} and \code{ChartColors}.}
\usage{
PrepareColors(dat, chart.type, scatter.colors.column = 4, palette = NULL,
  palette.custom.color = NULL, palette.custom.gradient.start = NULL,
  palette.custom.gradient.end = NULL, palette.custom.palette = NULL,
  fit.palette = NULL, fit.palette.custom.color = NULL,
  fit.palette.custom.gradient.start = NULL,
  fit.palette.custom.gradient.end = NULL, fit.palette.custom.palette = NULL,
  subslice.palette = NULL, subslice.palette.custom.color = NULL,
  subslice.palette.custom.gradient.start = NULL,
  subslice.palette.custom.gradient.end = NULL,
  subslice.palette.custom.palette = NULL)
}
\arguments{
\item{dat}{The data to be plotted, from which the number of required colors are deduced.}

\item{chart.type}{The name of plot to create.}

\item{scatter.colors.column}{For scatterplot data, the column of data which is used to determine the colors of each point.}

\item{palette}{Specifies the color vector to be used for the chart data. It can be (1) A named palette from grDevices, RColorBrewer colorspace, or colorRamps; or
(2) A vector of colors which will be recycled to the length of the number of colors needed; or
(3) one of \code{"Custom color"}, \code{"Custom gradient"} or \code{"Custom palette"}.
The last option gives the user greater control via additional parameters (see below).  If not specified, the colors used
are c("#5C9AD3", "#ED7D31", "#A5A5A5", "#FFC000", "#4473C5", "#70AD46",
"#255F91", "#9E480D", "#636365", "#987300", "#26408B", "#42682B")}

\item{palette.custom.color}{A single color provided as a hex or character string. Only used if \code{palette} is \code{"Custom color"}. The output vector will consist of \code{palette} repeated for the number of required colors.}

\item{palette.custom.gradient.start}{A color specifying the start of the gradient when \code{palette} is set to \code{"Custom gradient"}.}

\item{palette.custom.gradient.end}{A color specifying the end of the gradient when \code{palette} is set to \code{"Custom gradient"}.}

\item{palette.custom.palette}{A vector or comma separated list of colors which will be recycled to the desired length. Only used if \code{palette} is \code{"Custom palette"}.}

\item{fit.palette}{As per \code{palette} except for the chart fit lines. An additional option is \code{"Group colors"}, which uses the colors from \code{palette}.}

\item{fit.palette.custom.color}{As per \code{palette.custom.color}.}

\item{fit.palette.custom.gradient.start}{As per \code{palette.custom.gradient.start}.}

\item{fit.palette.custom.gradient.end}{As per \code{palette.custom.gradient.end}.}

\item{fit.palette.custom.palette}{As per \code{palette.custom.palette}.}

\item{subslice.palette}{As per \code{palette} except for pie chart subslice colors. An additional option is \code{"Group colors"}, which uses the colors from \code{palette}.}

\item{subslice.palette.custom.color}{As per \code{palette.custom.color}.}

\item{subslice.palette.custom.gradient.start}{As per \code{palette.custom.gradient.start}.}

\item{subslice.palette.custom.gradient.end}{As per \code{palette.custom.gradient.end}.}

\item{subslice.palette.custom.palette}{As per \code{palette.custom.palette}.}
}
\description{
Generates vectors of colors for use by CChart.
Wrapper for \code{GetNumColors} and \code{ChartColors}.
}
