% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparedata.R
\name{PrepareData}
\alias{PrepareData}
\title{Tidy Data For Use With Displayr Charting Functions}
\usage{
PrepareData(formChartType, subset = TRUE, weights = NULL,
  formTable = NULL, formTables = NULL, formBinary = NULL,
  pasted = list(NULL), raw.data = NULL, formTranspose = FALSE,
  number.format = list(NULL), missing = "Exclude cases with missing data",
  colors = list(NULL), row.names.to.remove = c("NET", "SUM"),
  col.names.to.remove = c("NET", "SUM"), show.labels = TRUE)
}
\arguments{
\item{formChartType}{character; chart type to be plotted}

\item{subset}{subset An optional vector specifying a subset of
observations to be used in the fitting process, or, the name of
a variable in \code{data}. It may not be an expression.}

\item{weights}{An optional vector of sampling weights, or, the name
of a variable in \code{data}. It may not be an expression.}

\item{formTable}{array; assumed to be a Qtable and will be
processed using \code{\link[flipTables]{AsBasicTable}}}

\item{formTables}{list of array; each component is assumed to be a
Qtable and will be processed using
\code{\link[flipTables]{AsBasicTable}}}

\item{formBinary}{a PickAny Multi Q variable}

\item{pasted}{list of length six; the first component of which is
assumed to be from a user-entered/pasted table; will be
processed by \code{\link{ParseUserEnteredTable}}}

\item{raw.data}{data.frame, containing Variables from a Q/Displayr
Data Set}

\item{formTranspose}{logical; should the supplied data be
transposed?}

\item{number.format}{list containing user-supplied parameters
relating to number formatting in the chart}

\item{missing}{character; One of \code{"Error if missing data"},
\code{"Exclude cases with missing data"} (the default, which is
equivalent to 'complete.cases'), and \code{"Use partial data"},
which removes no data; ignored if raw data is not supplied}

\item{colors}{list containing user-supplied values for colour
parameters in charts}

\item{row.names.to.remove}{character vector of row labels
specifying rows to remove from the returned table; default is
\code{c("NET", "SUM")}}

\item{col.names.to.remove}{character vector of column labels
specifying columns to remove from the returned table; default
is \code{c("NET", "SUM")}}

\item{show.labels}{logical; If \code{TRUE}, labels are used for
names in the data output if raw data is supplied}
}
\value{
If possible, a named vector or matrix, or if that is not
    posible or a data.frame is requested, a data.frame
}
\description{
For use before calling a Displayr charting function, ensures the
data to be used is in a convenient format for plotting
}
\details{
It is assumed that only one of \code{pasted},
    \code{formTable}, \code{formTables}, \code{formBinary},
    \code{raw.data} is non-NULL.  They are checked for nullity in
    that order
}
\seealso{
\code{\link[flipTables]{AsBasicTable}},
    \code{\link[flipData]{TidyRawData}},
    \code{\link[flipTransformations]{ParseUserEnteredTable}}
}
