% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparenumbers.R
\name{PrepareNumbers}
\alias{PrepareNumbers}
\title{Generates all d3 formats for use by CChart.
See https://github.com/d3/d3-format for more information on d3.
Wrapper for \code{ChartNumberFormat}.}
\usage{
PrepareNumbers(x.format.list = NULL, y.format.list = NULL,
  hover.format.list = NULL, data.labels.format.list = NULL)
}
\arguments{
\item{x.format.list}{A list of five unnamed items for formatting the x-axis as described in \code{\link{ChartNumberFormat}}.}

\item{y.format.list}{As per \code{x.format.list} except for y-axis formatting.}

\item{hover.format.list}{As per \code{x.format.list} except for hovertext formatting.}

\item{data.labels.format.list}{As per \code{x.format.list} except for data label formatting.}
}
\description{
Generates all d3 formats for use by CChart.
See https://github.com/d3/d3-format for more information on d3.
Wrapper for \code{ChartNumberFormat}.
}
