% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usetemplate.R
\name{print.AppearanceTemplate}
\alias{print.AppearanceTemplate}
\title{Print an appearance template object}
\usage{
\method{print}{AppearanceTemplate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{AppearanceTemplate}.}

\item{...}{Not used.}
}
\description{
Print an appearance template object
}
