% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartcolors.R
\name{ChartColors}
\alias{ChartColors}
\title{Generates a vector of colors}
\usage{
ChartColors(number.colors.needed, given.colors = qColors, reverse = FALSE,
  palette.start = 0, palette.end = 1, trim.light.colors = TRUE,
  custom.color = NA, custom.gradient.start = NA, custom.gradient.end = NA,
  custom.palette = NA)
}
\arguments{
\item{number.colors.needed}{The number of colors to generate.}

\item{given.colors}{Specifies the color vector to output. It can be (1) A named palette from grDevices, RColorBrewer colorspace, or colorRamps; (2) A vector of colors which will be recycled to length \code{number.colors.needed}; or (3) one of \code{"Custom color"}, \code{"Custom gradient"} or \code{"Custom palette"}. The last option gives the user greater control via additional parameters (see below).}

\item{reverse}{Whether the output color vector should be reversed. Ignored when \code{given.colors} is any of \code{"Custom color"}, \code{"Custom gradient"} or \code{"Custom palette"}.}

\item{palette.start}{A numeric in [0,1] specifying the start position of the palette}

\item{palette.end}{A numeric in [0,1] specifying the end position of the palette}

\item{trim.light.colors}{When selected, palette.start and palette.end will be set to remove light colors in the monochrome palettes (\code{"Blues","Greens","Greys","Oranges","Purples","Reds"}).}

\item{custom.color}{A single color provided as a hex or character string. Only used if \code{given.colors} is \code{"Custom color"}. The output vector will consist of \code{custom.color} repeated \code{number.colors.needed} (no interpolation).}

\item{custom.gradient.start}{A color specifying the start of the gradient when \code{given.colors} is set to \code{"Custom gradient"}.}

\item{custom.gradient.end}{A color specifying the end of the gradient when \code{given.colors} is set to \code{"Custom gradient"}.}

\item{custom.palette}{A vector or comma separated list of colors which will be recycled to the desired length. Only used if \code{given.colors} is \code{"Custom palette"}.}
}
\description{
Generates a vector of colors from a palette or vector of colors. 
Alpha channels are ignored and set to 255/FF.
}
\examples{
ChartColors(5, given.colors = c("blue", "orange", "green"))
ChartColors(5, given.colors = "blue")
ChartColors(5, given.colors = "#9CFF73")
ChartColors(5, given.colors = "Set3", reverse = TRUE)
plot(1:10,1:10,pch=19, col=ChartColors(10,"Reds",trim.light.colors=TRUE, reverse=TRUE))
}
