% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartnumberformat.R
\name{ChartNumberFormat}
\alias{ChartNumberFormat}
\title{Generates a d3 number format for use by charting functions.}
\usage{
ChartNumberFormat(number.format, as.percentages = FALSE)
}
\arguments{
\item{number.format}{A list of five unnamed items in the following order:
1) The type of number formatting. One of \code{"Automatic"}, \code{"Category"}, \code{"Number"},
 \code{"Percentage"}, \code{"Date/Time"}, \code{"Currency"}, \code{"Metric units suffix"}, 
 \code{"Scientific"} and \code{"Custom"}.
2) A d3 date format. If specified, this will be used provided type is neither \code{"Automatic"}
nor \code{"Category"} nor a custom format is not specified.
3) A custom d3 number format. If specifed this will be used in preference to all other inputs
apart from \code{"Automatic"} and \code{"Category"}.
4) A boolean indicating whether to separate thousands with commas.
5) An integer specifying the number of decimal places, or for \code{"Metric units suffix"}, the number
of significant digits.}

\item{as.percentages}{Whether the \code{"Automatic"} formatting should be as percentages.}
}
\description{
See https://github.com/d3/d3-format for more information on d3.
}
\details{
Defaults to \code{"Automatic"} if no type is specified.
}
