% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usetemplate.R
\name{GetBrandColors}
\alias{GetBrandColors}
\title{Identify colors to be used with brands in the data}
\usage{
GetBrandColors(
  template,
  input.data = NULL,
  filter = NULL,
  chart.type = "",
  scatter.colors.column = 4,
  multi.color.series = FALSE,
  brand.names = NULL
)
}
\arguments{
\item{template}{A list specifying color palettes and other visualization options.}

\item{input.data}{Input data for the visualization. The is usually a vector or table. 
It should be normalized from \code{flipChart::PrepareData}.}

\item{filter}{Optional filter which can be used with the input data. The label of the filter is used when appropriate.}

\item{chart.type}{The visualization which will be applied to the data.}

\item{scatter.colors.column}{For scatter plots, an integer specifying the data column used to specify the colors.}

\item{multi.color.series}{For bar and column charts, a logical indicating how colors are used.}

\item{brand.names}{If a vector of names is provided then only \code{template} is used.}
}
\description{
Identify colors to be used with brands in the data
}
