% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usetemplate.R
\name{GetBrandsFromData}
\alias{GetBrandsFromData}
\title{Identify brand names from the input data}
\usage{
GetBrandsFromData(data, filter, chart.type, scatter.colors.column = 4,
  multi.color.series = FALSE)
}
\arguments{
\item{data}{Input data for the visualization. The is usually a vector or table. It should be normalized from \code{flipChart::PrepareData}.}

\item{filter}{Optional filter which can be used with the input data. The label of the filter is used when appropriate.}

\item{chart.type}{The visualization which will be applied to the data.}

\item{scatter.colors.column}{For scatter plots, an integer specifying the data column used to specify the colors.}

\item{multi.color.series}{For bar and column charts, a logical indicating how colors are used.}
}
\description{
Extracts the names of the data series from the input data, to be associated with different colors.
This usually corresponds to columns in the input table. But for charts such as 'Pyramid' or 'Pie', each 
entry in the vector (or each row in the table) corresponds to a new data series.
}
