% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usetemplate.R
\name{GetPalette}
\alias{GetPalette}
\title{Determine colors to be used in a visualization}
\usage{
GetPalette(palette, template)
}
\arguments{
\item{palette}{A string naming the palette to be used.}

\item{template}{A list specifying color palettes and other yyvisualization options.}
}
\description{
Determine colors to be used in a visualization
}
