% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getvectorofcolors.R
\name{GetVectorOfColors}
\alias{GetVectorOfColors}
\title{Identify colors to be used with brands in the data}
\usage{
GetVectorOfColors(
  template,
  input.data,
  filter,
  chart.type,
  scatter.colors.column = FALSE,
  multi.color.series = FALSE,
  palette = NULL,
  palette.custom.color = NULL,
  palette.custom.gradient.start = NULL,
  palette.custom.gradient.end = NULL,
  palette.custom.palette = NULL,
  type = "Series"
)
}
\arguments{
\item{template}{A list specifying color palettes and other visualization options.}

\item{input.data}{Input data for the visualization. The is usually a vector or table. 
It should be normalized from \code{flipChart::PrepareData}.}

\item{filter}{Optional filter which can be used with the input data. The label of the filter is used when appropriate.}

\item{chart.type}{The visualization which will be applied to the data.}

\item{scatter.colors.column}{For scatter plots, an integer specifying the data column used to specify the colors.}

\item{multi.color.series}{For bar and column charts, a logical indicating how colors are used.}

\item{palette}{Specifies the color vector to output. It can be;
(1) A named palette from grDevices, RColorBrewer colorspace, or colorRamps;
(2) A vector of colors which will be recycled to length \code{number.colors.needed}; or
(3) one of \code{"Custom color"}, \code{"Custom gradient"} or \code{"Custom palette"}.
The last option gives the user greater control via additional parameters (see below).  If not specified, the colors used
are c("#5C9AD3", "#ED7D31", "#A5A5A5", "#FFC000", "#4473C5", "#70AD46",
            "#255F91", "#9E480D", "#636365", "#987300", "#26408B", "#42682B")}

\item{palette.custom.color}{A single color provided as a hex or character string.
Only used if \code{palette} is \code{"Custom color"}. The output vector will
consist of \code{custom.color} repeated to the desired length 
(no interpolation).}

\item{palette.custom.gradient.start}{A color specifying the start of 
the gradient when \code{palette} is set to \code{"Custom gradient"}.}

\item{palette.custom.gradient.end}{A color specifying the end of the 
gradient when \code{palette} is set to \code{"Custom gradient"}.}

\item{palette.custom.palette}{A vector or comma separated list of colors 
which will be recycled to the desired length.}

\item{type}{Describes the type of data which the color vector will be applied to.
One of "Series" or "Pie subslice".}
}
\description{
Identify colors to be used with brands in the data
}
