% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartcolors.R
\name{StripAlphaChannel}
\alias{StripAlphaChannel}
\title{Removes the last two characters from a vector of hexadecimal colors
that include an alpha channel.}
\usage{
StripAlphaChannel(hex.colors)
}
\arguments{
\item{hex.colors}{A character vector of colors in hexadeximal format.}
}
\description{
Removes the last two characters from a vector of hexadecimal colors
that include an alpha channel.
}
\examples{
c <- c("#FFFFFFFF", "#ABCDEF3D")
StripAlphaChannel(c)
}

