% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartcolors.R
\name{StripAlphaChannel}
\alias{StripAlphaChannel}
\title{Removes the last two characters from a vector of hexadecimal colors
that include an alpha channel.}
\usage{
StripAlphaChannel(hex.colors)
}
\arguments{
\item{hex.colors;}{a vector of colors in hexadeximal format.}

\item{given.colors}{Character; a vector containing one or more named
colors from grDevices OR one or more specified hex value colors OR a single
named palette from grDevices, RColorBrewer, colorspace, or colorRamps.}

\item{reverse}{Logical; if the output color vector shour be reversed.}
}
\description{
Removes the last two characters from a vector of hexadecimal colors
that include an alpha channel.
}
\examples{
c <- c("#FFFFFFFF", "#ABCDEF3D")
StripAlphaChannel(c)
}

